Quarter_03/Class_Challanges/27-Apr-2025/chainlit_UI/pyproject.toml
[project]
name = "chainlit-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.13",
]
